+++ super_native_extensions/rust/src/darwin/ios/reader.rs
@@ -207,7 +207,7 @@ impl PlatformDataReader {
         Ok(res)
     }
 
-    pub fn item_format_is_synthetized(
+    pub fn item_format_is_synthesized(
         &self,
         _item: i64,
         _format: &str,